requirements.txt
Django<1.7
WebOb==1.3.1
WebTest==2.0.15
argparse==1.2.1
beautifulsoup4==4.3.2
django-debug-toolbar==1.1
django-dynamic-fixture==1.7.0
django-webtest==1.7.7
gunicorn==18.0
oauthlib==0.6.1
python-openid==2.2.5
python-social-auth==0.1.23
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
sqlparse==0.1.11
waitress==0.8.8
wsgiref==0.1.2
mock==1.0.1
psycopg2==2.5.2
South==0.8.4
django-pipeline==1.3.24
factory-boy==2.3.1
django-extensions==1.3.7
Markdown==2.4
diff-match-patch
